# client/src/api/openapi.json
{"openapi":"3.0.0","paths":{"/vehicles":{"post":{"operationId":"VehiclesController_createVehicle","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVehicleDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleResponseDto"}}}}},"summary":"Create a new vehicle","tags":["Vehicles"]},"get":{"operationId":"VehiclesController_listVehicles","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VehicleResponseDto"}}}}}},"summary":"List all vehicles","tags":["Vehicles"]}},"/vehicles/{licensePlate}":{"get":{"operationId":"VehiclesController_getVehicleByLicensePlate","parameters":[{"name":"licensePlate","required":true,"in":"path","description":"Vehicle license plate","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleResponseDto"}}}}},"summary":"Get a vehicle by license plate","tags":["Vehicles"]}},"/vehicles/{licensePlate}/summary":{"get":{"operationId":"VehiclesController_getVehicleSummary","parameters":[{"name":"licensePlate","required":true,"in":"path","description":"Vehicle license plate","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleSummaryDto"}}}}},"summary":"Get aggregated trip summary for a vehicle","tags":["Vehicles"]}},"/vehicles/{licensePlate}/trip":{"post":{"operationId":"TripsController_createTrip","parameters":[{"name":"licensePlate","required":true,"in":"path","description":"Vehicle license plate","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTripDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripResponseDto"}}}}},"summary":"Record a new trip for a vehicle","tags":["Trips"]}},"/trips":{"get":{"operationId":"TripsController_listTrips","parameters":[{"name":"licensePlate","required":false,"in":"query","description":"Filter by vehicle license plate","schema":{"type":"string"}},{"name":"startDate","required":false,"in":"query","description":"Return trips starting on or after this date (ISO 8601)","schema":{"format":"date-time","example":"2024-01-01T00:00:00Z","type":"string"}},{"name":"endDate","required":false,"in":"query","description":"Return trips starting on or before this date (ISO 8601)","schema":{"format":"date-time","example":"2024-12-31T23:59:59Z","type":"string"}},{"name":"page","required":false,"in":"query","description":"Page number (default: 1)","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Items per page, max 100 (default: 20)","schema":{"example":20,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedTripsDto"}}}}},"summary":"List trips with optional filters and pagination","tags":["Trips"]}}},"info":{"title":"Voyage Fleet API","description":"REST API for recording and analysing fleet vehicle trips","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateVehicleDto":{"type":"object","properties":{"name":{"type":"string","description":"Display name for the vehicle","example":"Israel Express"},"licensePlate":{"type":"string","description":"Unique licence plate number","example":"IL-EX-001"}},"required":["name","licensePlate"]},"VehicleResponseDto":{"type":"object","properties":{"name":{"type":"string","description":"Display name for the vehicle","example":"Israel Express"},"licensePlate":{"type":"string","description":"Unique licence plate number","example":"IL-EX-001"},"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"}},"required":["name","licensePlate","id","createdAt"]},"VehicleSummaryDto":{"type":"object","properties":{"vehicle":{"$ref":"#/components/schemas/VehicleResponseDto"},"tripCount":{"type":"number"},"totalDistanceKm":{"type":"number"},"totalFuelConsumed":{"type":"number"},"averageDurationMinutes":{"type":"number"}},"required":["vehicle","tripCount","totalDistanceKm","totalFuelConsumed","averageDurationMinutes"]},"CreateTripDto":{"type":"object","properties":{"startedAt":{"format":"date-time","type":"string","description":"Trip start time (ISO 8601)","example":"2024-06-01T08:00:00Z"},"endedAt":{"format":"date-time","type":"string","description":"Trip end time (ISO 8601)","example":"2024-06-01T09:30:00Z"},"distanceKm":{"type":"number","description":"Distance covered in kilometres","example":145.5},"fuelConsumed":{"type":"number","description":"Fuel or energy consumed (litres / kWh)","example":18.3}},"required":["startedAt","endedAt","distanceKm","fuelConsumed"]},"TripResponseDto":{"type":"object","properties":{"startedAt":{"format":"date-time","type":"string","description":"Trip start time (ISO 8601)","example":"2024-06-01T08:00:00Z"},"endedAt":{"format":"date-time","type":"string","description":"Trip end time (ISO 8601)","example":"2024-06-01T09:30:00Z"},"distanceKm":{"type":"number","description":"Distance covered in kilometres","example":145.5},"fuelConsumed":{"type":"number","description":"Fuel or energy consumed (litres / kWh)","example":18.3},"tripId":{"type":"string"},"vehicleId":{"type":"string"},"durationMinutes":{"type":"number"},"createdAt":{"format":"date-time","type":"string"}},"required":["startedAt","endedAt","distanceKm","fuelConsumed","tripId","vehicleId","durationMinutes","createdAt"]},"PaginatedTripsDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/TripResponseDto"}},"totalTrips":{"type":"number"},"page":{"type":"number"},"limit":{"type":"number"}},"required":["data","totalTrips","page","limit"]}}}}